requirements.txt
fastapi
uvicorn
pre-commit
ruff
pytest
pytest-cov
httpx
sqlalchemy
requests
python-dotenv
pydantic
